+++ src/VaultSharp/V1/SecretsEngines/GoogleCloud/GoogleCloudOAuth2Token.cs
@@ -1,0 +1,19 @@
+﻿using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.GoogleCloud
+{
+    /// <summary>
+    /// Represents the GoogleCloud OAuth2 Token.
+    /// </summary>
+    public class GoogleCloudOAuth2Token
+    {
+        /// <summary>
+        /// Gets or sets the token.
+        /// </summary>
+        /// <value>
+        /// The token.
+        /// </value>
+        [JsonProperty("token")]
+        public string Token { get; set; }
+    }
+}