+++ src/VaultSharp/V1/SecretsEngines/SecretsEngineDefaultPaths.cs
@@ -9,6 +9,7 @@
         public const string Consul = "consul";
         public const string Cubbyhole = "cubbyhole";
         public const string Database = "database";
+        public const string GoogleCloud = "gcp";
         public const string KeyValue = "kv";
         public const string Identity = "identity";
         public const string Nomad = "nomad";