+++ src/VaultSharp/V1/SecretsEngines/SecretsEngineProvider.cs
@@ -5,6 +5,7 @@ using VaultSharp.V1.SecretsEngines.AWS;
 using VaultSharp.V1.SecretsEngines.Consul;
 using VaultSharp.V1.SecretsEngines.Cubbyhole;
 using VaultSharp.V1.SecretsEngines.Database;
+using VaultSharp.V1.SecretsEngines.GoogleCloud;
 using VaultSharp.V1.SecretsEngines.Identity;
 using VaultSharp.V1.SecretsEngines.KeyValue;
 using VaultSharp.V1.SecretsEngines.PKI;
@@ -28,6 +29,7 @@ namespace VaultSharp.V1.SecretsEngines
             Consul = new ConsulSecretsEngineProvider(polymath);
             Cubbyhole = new CubbyholeSecretsEngineProvider(polymath);
             Database = new DatabaseSecretsEngineProvider(polymath);
+            GoogleCloud = new GoogleCloudSecretsEngineProvider(polymath);
 
             KeyValue = new KeyValueSecretsEngineProvider(polymath);
             PKI = new PKISecretsEngineProvider(polymath);
@@ -43,6 +45,8 @@ namespace VaultSharp.V1.SecretsEngines
         public ICubbyholeSecretsEngine Cubbyhole { get; }
 
         public IDatabaseSecretsEngine Database { get; }
+
+        public IGoogleCloudSecretsEngine GoogleCloud { get; }
 
         public IKeyValueSecretsEngine KeyValue { get; }