+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   * AWS Secrets Engine: Add support for generating dynamic IAM credentials & STS IAM credentials.
   * Cubbhole Secrets Engine: Add support for read secret, read paths, write secret and delete secret APIs.
   * Database Secrets Engine: Add support for generating dynamic DB credentials.
+  * GoogleCloud Secrets Engine: Add support for generating OAuth2 Token & Service Account Key.
 
 ## 0.10.4001 (August 10, 2018)